requirements.txt
Django==1.7.4
Pillow==2.7.0
South==1.0
Unidecode==0.04.17
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-ganalytics==0.4
django-compressor==1.4
django-database-storage==0.1.2
django-libsass==0.2
django-modelcluster==0.5
django-taggit==0.12.2
django-toolbelt==0.0.1
django-treebeard==2.0
django-widget-tweaks==1.3
gunicorn==19.2.1
html5lib==0.999
libsass==0.6.2
psycopg2==2.6
requests==2.5.1
six==1.9.0
static3==0.5.1
unicodecsv==0.9.4
wagtail==0.8.4
wsgiref==0.1.2
